spotless-implementation

+++ app/src/main/java/phone/vishnu/quotes/activity/SplashActivity.java
@@ -1,3 +1,22 @@
+/*
+ * Copyright (C) 2019 - 2019-2021 Vishnu Sanal. T
+ *
+ * This file is part of Quotes Status Creator.
+ *
+ * Quotes Status Creator is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 package phone.vishnu.quotes.activity;
 
 import android.content.Intent;
@@ -7,14 +26,11 @@ import android.os.Environment;
 import android.os.Handler;
 import android.view.WindowManager;
 import android.widget.TextView;
-
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.app.AppCompatDelegate;
-
 import java.io.File;
 import java.util.ArrayList;
 import java.util.Objects;
-
 import phone.vishnu.quotes.R;
 import phone.vishnu.quotes.fragment.TourFragment;
 import phone.vishnu.quotes.helper.SharedPreferenceHelper;
@@ -40,9 +56,7 @@ public class SplashActivity extends AppCompatActivity {
 
         if (sharedPreferenceHelper.isNewFirstRun() || sharedPreferenceHelper.isFirstRun())
             showNewTour();
-        else
-            initTasks();
-
+        else initTasks();
     }
 
     private void checkTheme(int i) {
@@ -56,8 +70,11 @@ public class SplashActivity extends AppCompatActivity {
     private void removeFonts() {
         sharedPreferenceHelper.deleteFontPreference();
 
-        File root = new File(Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOCUMENTS),
-                getString(R.string.app_name));
+        File root =
+                new File(
+                        Environment.getExternalStoragePublicDirectory(
+                                Environment.DIRECTORY_DOCUMENTS),
+                        getString(R.string.app_name));
 
         File[] files = root.listFiles();
 
@@ -66,20 +83,23 @@ public class SplashActivity extends AppCompatActivity {
         if (files != null) {
             for (File file : files) {
                 if (file.getAbsolutePath().endsWith(".ttf")) {
-                    if (arrayList.contains(file.getName()))
-                        file.delete();
+                    if (arrayList.contains(file.getName())) file.delete();
                 }
             }
         }
     }
 
     private void showNewTour() {
-        getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN, WindowManager.LayoutParams.FLAG_FULLSCREEN);
+        getWindow()
+                .setFlags(
+                        WindowManager.LayoutParams.FLAG_FULLSCREEN,
+                        WindowManager.LayoutParams.FLAG_FULLSCREEN);
 
-        getSupportFragmentManager().beginTransaction()
+        getSupportFragmentManager()
+                .beginTransaction()
                 .replace(R.id.splashScreenConstraintLayout, TourFragment.newInstance())
-                .addToBackStack(null).commit();
-
+                .addToBackStack(null)
+                .commit();
     }
 
     public void moveToNext() {
@@ -93,7 +113,8 @@ public class SplashActivity extends AppCompatActivity {
 
         if (colorString.equals("#00000000")) colorString = "#607D8B";
 
-        ((TextView) findViewById(R.id.splashScreenAppNameTextView)).setTextColor(Color.parseColor(colorString));
+        ((TextView) findViewById(R.id.splashScreenAppNameTextView))
+                .setTextColor(Color.parseColor(colorString));
 
         int SPLASH_TIMEOUT = 1;
         new Handler().postDelayed(this::moveToNext, SPLASH_TIMEOUT * 1000);
